Added PlayerClient and LocalPlay game class
PlayerClient has been added in order to connect with the server to play
games etc.
Local PlayGame has been added so that data can be collected and sent to
server, and also it acts as a means of displaying the information so the
user can select appropriate options etc
Tested server, setupclient, and playerclient - can create and play a
game, access scores, view all quizzes.
Need to address client ID and Quiz ID for when recalling from file.